Rename HttpServer to ServerPrefix, and do not listen on it directly
Rename HttpServer to ServerPrefix, which is closer to the terminology
used in https://datatracker.ietf.org/doc/html/rfc6962#section-4 and in
https://github.com/C2SP/C2SP/blob/main/static-ct-api.md#checkpoints to
describe the prefix of the CA server. In the future, this can be used as
an identifier for the server (e.g., as a checkpoint origin line).

Also do not listen on ServerPrefix directly with the 'mtc ca serve'
subcommand, since it may not be in 'host:port' format. The `listen-addr`
flag provides the listening address for the server, and we assume that
it's behind a reverse proxy that ensures that requests to ServerPrefix
are routed correctly.

Closes #22