Add compact umbilical format (#31)
Add compact umbilical format

This is only available for CAs that require umbilical evidence.

Instead of storing the certificates directly in the evidence, we store
the sha256 fingerprint.

In each batch, the CA publishes an extra file called umbilical-certificates.
This is just a list of X509 certificates together with an index to look
them up by hash. The certificate referred to in the evidence doesn't
have to appear in the same batch, but could appear in any batch within
the validity window.

This deduplicates intermediates, and deduplicates when multiple MTCs are
issued using the same umbilical leaf certificate.

Co-authored-by: Luke Valenta <lvalenta@cloudflare.com>