Add a method for checking validity window coverage
Define a method `Verify()` for checking that a certificate is covered by
a given validity window. It also checks whether the certificate has
expired or the issuer matches the CA. The API is designed to accommodate
any additional checks we might need in the future.

The method assumes the caller has already verified the validity window's
signature, e.g., using `(w *SignedValidityWindow) UnmarshalBinary(data
[]byte, p *CAParams)`.

To facilitate testing, add a method to the CA for getting a signed
validity window.

Finally, add some documentation and fix a typo.